Work on layout animation updater types
