+++ src/reanimated2/commonTypes.ts
@@ -1,0 +1,9 @@
+import { ViewStyle, TextStyle } from 'react-native';
+import { Animation } from './animation/commonTypes';
+
+export type StyleProps =
+  | ViewStyle
+  | TextStyle
+  | { originX?: number; originY?: number };
+
+export type AnimatedStyle = Record<string, Animation>;
