+++ src/reanimated2/layoutReanimation/AnimatedRoot.tsx
@@ -1,34 +1,6 @@
 /* global _stopObservingProgress, _startObservingProgress */
-import { Platform, requireNativeComponent } from 'react-native';
-import React from 'react';
 import { runOnUI } from '../core';
 import { withStyleAnimation } from '../animation/styleAnimation';
-
-let REALayoutView: any;
-let meabyREALayoutView: any;
-if (Platform.OS === 'web' && !requireNativeComponent) {
-  meabyREALayoutView = React.Component;
-} else {
-  try {
-    meabyREALayoutView = (requireNativeComponent(
-      'REALayoutView'
-    ) as unknown) as React.Component;
-  } catch (e) {
-    // in dev mode after liver reload, it is tried to register the object again and it crash app
-    if (!__DEV__) {
-      throw e;
-    }
-  }
-  if (meabyREALayoutView) {
-    REALayoutView = meabyREALayoutView;
-  } // else use last component
-}
-
-export class AnimatedLayout extends React.Component {
-  render(): React.ReactElement {
-    return <REALayoutView collapsable={false} {...this.props} />;
-  }
-}
 
 // Register LayoutAnimationRepository
