+++ src/reanimated2/layoutReanimation/index.ts
@@ -1,3 +1,4 @@
-export * from './AnimatedRoot';
+require('./AnimatedRoot');
+export * from './AnimatedLayout';
 export * from './animationBuilder';
 export * from './defaultAnimations';
